fixes for bug found by Thomas Ponten Sicheritz. When doing blastall with gapped alignment (-g=F), the descriptions will include an extra term N that indicates the number of alignments for the same subject.                                                                Score     E Sequences producing significant alignments:                    (bits)  Value  N
This is also reported in the alignments as "Expect(??) = XXXX".

Thus, I have added a num_alignments member to the Record.HSP and
Record.Description classes.  I made changes to the scanners in
NCBIWWW and NCBIStandalone and the appropriate consumers in NCBIStandalone
to parse this information.

Plus, I added a new event called "description_header" so that I will
know whether to expect the "N" term in the descriptions.

I updated the docs and regression tests accordingly.